+++ README.md
@@ -5,8 +5,18 @@
 ###### Install node modules:
 `npm install`
 
+###### Create environment variables file and enter API credentials
+
+- Filename: `.env`
+- Contents:
+```
+MERCHANT_ID = your_merchant_id
+PUBLIC_KEY = your_public_api_key
+PRIVATE_KEY = your_private_api_key
+```
+
 ###### Start app: 
 `npm start`
 
-###### In the browser, go to:
-[`localhost:8000`](http://localhost:8000/)
+###### In the browser, got to:
+`localhost:8000`
