+++ config/gateway.js
@@ -3,8 +3,11 @@ import dotenv from "dotenv";
 
 dotenv.config();
 
-// <--- SET UP GATEWAY CONFIGURATION BELOW --->
-var gateway =  "Enter API credentials here"
-// <--- / END --->
+var gateway = new braintree.BraintreeGateway({
+  environment: braintree.Environment.Sandbox,
+  merchantId: process.env.MERCHANT_ID,
+  publicKey: process.env.PUBLIC_KEY,
+  privateKey: process.env.PRIVATE_KEY,
+});
 
 export default gateway;
