--- controllers/CLIENT_TOKEN.js
@@ -1,9 +1,0 @@
-import gateway from "../config/gateway.js";
-
-export const clientTokenGeneration = async () => {
-
-};
-
-export const clientTokenForDropIn = (req, res) => {
-  res.render("DROP-IN")
-};
