--- controllers/TRANSACTION.js
@@ -1,6 +1,0 @@
-import gateway from "../config/gateway.js";
-
-export const transactionSale = (req, res) => {
-
-};
-
