--- data/data.json
@@ -1,1 +1,0 @@
-{"name":"","credentialsCompleted":true,"clientTokenCompleted":true,"dropinCompleted":true,"transactionCompleted":false}
