+++ package.json
@@ -14,9 +14,8 @@
   "dependencies": {
     "braintree": "^3.15.0",
     "dotenv": "^8.2.0",
-    "ejs": "^3.1.6",
     "express": "^4.18.2",
-    "fs": "^0.0.1-security",
+    "ejs": "^3.1.6",
     "node-sass": "^9.0.0",
     "nodemon": "^2.0.22",
     "path": "^0.12.7"
