+++ public/hostedFields.css
@@ -3,20 +3,32 @@
   min-height: 100vh;
   width: 100%;
   padding: 2%;
+  display: -webkit-box;
+  display: -ms-flexbox;
   display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
 }
+
 #hostedFields form {
   width: 45%;
 }
 
 #cardForm {
   background-color: white;
-  box-shadow: 3px 3px 3px 3px gray;
+  -webkit-box-shadow: 3px 3px 3px 3px gray;
+          box-shadow: 3px 3px 3px 3px gray;
   padding: 3%;
 }
+
 #cardForm label {
   margin-bottom: 1%;
   font-weight: bold;
@@ -34,6 +46,7 @@
   height: 35px;
   padding: 1%;
   font-size: 18px;
+  -webkit-transition: border-color 160ms;
   transition: border-color 160ms;
 }
 
@@ -62,4 +75,5 @@
   position: absolute;
   top: 2%;
   right: 2%;
-}/*# sourceMappingURL=hostedFields.css.map */
+}
+/*# sourceMappingURL=hostedFields.css.map */
