+++ public/style.css
@@ -1,114 +1,89 @@
 #homePage {
   text-align: center;
-  min-height: 100vh;
+  height: 100vh;
   width: 100%;
   background-color: #020213;
   padding: 1%;
 }
+
 #homePage h1 {
-  height: 10vh;
-}
-#homePage p, #homePage button {
-  font-family: "Raleway", sans-serif;
-}
-#homePage h1,
-#homePage h2,
-#homePage h5 {
   color: #F4F4F9;
+  display: -webkit-box;
+  display: -ms-flexbox;
   display: flex;
-  justify-content: center;
-  align-items: center;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
   height: 10vh;
   text-decoration: none;
-  font-family: "Raleway", sans-serif;
+  font-family: 'Raleway', sans-serif;
 }
-#homePage h5 {
-  margin: -1%;
-}
+
 #homePage hr {
   color: #F4F4F9;
   width: 75%;
   margin: 2% auto;
 }
+
 #homePage main section {
   margin: auto;
   width: 80%;
+  display: -ms-grid;
   display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(30%, 1fr));
-  -moz-column-gap: 20px;
-       column-gap: 20px;
+  -ms-grid-columns: (minmax(30%, 1fr))[auto-fit];
+      grid-template-columns: repeat(auto-fit, minmax(30%, 1fr));
+  -webkit-column-gap: 20px;
+          column-gap: 20px;
   row-gap: 20px;
 }
+
 #homePage main section a {
+  display: -webkit-box;
+  display: -ms-flexbox;
   display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 20vh;
-  min-width: 20%;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  min-height: 10vh;
   font-size: 1.5rem;
-}
-
-#homeButton {
-  position: absolute;
-  top: 5%;
-  left: 5%;
-}
-#homeButton i {
-  font-size: xx-large;
-}
-
-#locked {
-  pointer-events: none;
-  filter: grayscale(1);
-}
-
-#tokenContainer {
-  flex-direction: column;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-#tokenContainer p {
-  color: white;
-  font-size: large;
-}
-#tokenContainer p button {
-  color: blue;
-  text-decoration: none;
-  outline-style: none;
-  background: transparent;
-  border: none;
-  padding: 0;
-  margin: 0;
-}
-#tokenContainer code, #tokenContainer pre {
-  background-color: black;
-  color: lime;
-  letter-spacing: 2px;
-  border: 1px solid #999;
-  display: block;
-  padding: 2%;
-  width: 90%;
-  text-align: left;
 }
 
 #checkout-page {
   height: 100vh;
-  width: 60%;
+  width: 100%;
   background-color: #F5F5F5;
   padding: 1%;
+  display: -webkit-box;
+  display: -ms-flexbox;
   display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
+  -webkit-box-align: center;
+      -ms-flex-align: center;
+          align-items: center;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
 }
+
 #checkout-page table {
+  width: -webkit-fit-content;
   width: -moz-fit-content;
   width: fit-content;
 }
+
 #checkout-page form {
   width: 55%;
 }
+
 #checkout-page form label {
   font-weight: bold;
   margin-bottom: 1%;
@@ -116,9 +91,11 @@
 
 #cardForm {
   background-color: white;
-  box-shadow: 2px 2px 2px 2px gray;
+  -webkit-box-shadow: 2px 2px 2px 2px gray;
+          box-shadow: 2px 2px 2px 2px gray;
   padding: 2%;
 }
+
 #cardForm label {
   font-weight: bold;
 }
@@ -126,14 +103,23 @@
 #dropin-holder {
   height: 45vh;
   width: 100%;
+  display: -webkit-box;
+  display: -ms-flexbox;
   display: flex;
-  justify-content: center;
-  flex-direction: column;
+  -webkit-box-pack: center;
+      -ms-flex-pack: center;
+          justify-content: center;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+      -ms-flex-direction: column;
+          flex-direction: column;
 }
 
 .docs-card {
   background-color: white;
   margin-top: 4%;
-  box-shadow: 4px 4px 4px 4px silver;
+  -webkit-box-shadow: 4px 4px 4px 4px silver;
+          box-shadow: 4px 4px 4px 4px silver;
   padding: 1%;
-}/*# sourceMappingURL=style.css.map */
+}
+/*# sourceMappingURL=style.css.map */
