+++ public/style.scss
@@ -1,149 +1,84 @@
 #homePage {
-  text-align: center;
-  min-height: 100vh;
-  width: 100%;
-  background-color: #020213;
-  padding: 1%;
-
-  h1 {
-    height: 10vh;
-  }
-
-  p, button {
-    font-family: 'Raleway', sans-serif;
-  }
-
-  h1,
-  h2,
-  h5 {
-    color: #F4F4F9;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 10vh;
-    text-decoration: none;
-    font-family: 'Raleway', sans-serif;
-  }
-
-  h5 {
-    margin: -1%;
-  }
-
-  hr {
-    color: #F4F4F9;
-    width: 75%;
-    margin: 2% auto;
-  }
-
-  main {
-    section {
-      margin: auto;
-      width: 80%;
-      display: grid;
-      grid-template-columns: repeat(auto-fit, minmax(30%, 1fr));
-      column-gap: 20px;
-      row-gap: 20px;
-      a {
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        min-height: 20vh;
-        min-width: 20%;
-        font-size: 1.5rem;
+    text-align: center;
+    height: 100vh;
+    width: 100%;
+    background-color: #020213;
+    padding: 1%;
+    h1 {
+      color: #F4F4F9;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      height: 10vh;
+      text-decoration: none;
+      font-family: 'Raleway', sans-serif;
+    }
+    hr {
+        color: #F4F4F9;
+        width: 75%;
+        margin: 2% auto;
+    }
+    main {
+      section {
+        margin: auto;
+        width: 80%;
+        display: grid;
+        grid-template-columns: repeat(auto-fit, minmax(30%, 1fr));
+        column-gap: 20px;
+        row-gap: 20px;
+        a {
+          display: flex;
+          justify-content: center;
+          align-items: center;
+          min-height: 10vh;
+          font-size: 1.5rem;
+        }
       }
     }
   }
-}
-
-#homeButton {
-  position: absolute;
-  top: 5%;
-  left: 5%;
-  i {
-    font-size: xx-large;
-  }
-}
-
-#locked {
-  pointer-events: none;
-  filter: grayscale(1);
-}
-
-#tokenContainer {
-  flex-direction: column;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  p {
-    color: white;
-    font-size: large;
-    button {
-      color: blue;
-      text-decoration: none;
-      outline-style: none;
-      background: transparent;
-      border: none;
-      padding: 0;
-      margin: 0;
+  
+  #checkout-page {
+    height: 100vh;
+    width: 100%;
+    background-color: #F5F5F5;
+    padding: 1%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    table {
+      width: fit-content;
+    }
+    form {
+      width: 55%;
+      label {
+        font-weight: bold;
+        margin-bottom: 1%;
+      }
     }
   }
-  code, pre {
-    background-color: black;
-    color: lime;
-    letter-spacing: 2px;
-    border: 1px solid #999;
-    display: block;
+  
+  #cardForm {
+    background-color: white;
+    box-shadow: 2px 2px 2px 2px gray;
     padding: 2%;
-    width: 90%;
-    text-align: left;
-  }
-}
-
-#checkout-page {
-  height: 100vh;
-  width: 60%;
-  background-color: #F5F5F5;
-  padding: 1%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-
-  table {
-    width: fit-content;
-  }
-
-  form {
-    width: 55%;
-
     label {
       font-weight: bold;
-      margin-bottom: 1%;
     }
   }
-}
 
-#cardForm {
-  background-color: white;
-  box-shadow: 2px 2px 2px 2px gray;
-  padding: 2%;
-
-  label {
-    font-weight: bold;
+  #dropin-holder{
+    height: 45vh;
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    flex-direction: column;
   }
-}
 
-#dropin-holder {
-  height: 45vh;
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-}
-
-.docs-card {
-  background-color: white;
-  margin-top: 4%;
-  box-shadow: 4px 4px 4px 4px silver;
-  padding: 1%;
-}
+  .docs-card {
+    background-color: white;
+    margin-top: 4%;
+    box-shadow: 4px 4px 4px 4px silver;
+    padding: 1%;
+  }
+
