--- public/style2.css
@@ -1,79 +1,0 @@
-#homePage {
-  text-align: center;
-  height: 100vh;
-  width: 100%;
-}
-
-#homePage h1 {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: 10vh;
-  text-decoration: underline;
-}
-
-#homePage main section {
-  margin: auto;
-  width: 80%;
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(30%, 1fr));
-  -moz-column-gap: 20px;
-       column-gap: 20px;
-  row-gap: 20px;
-}
-
-#homePage main section a {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 10vh;
-}
-
-#checkout-page {
-  padding: 3%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-}
-
-#checkout-page table {
-  width: -moz-fit-content;
-  width: fit-content;
-}
-
-#checkout-page form {
-  width: 55%;
-}
-
-#checkout-page form label {
-  font-weight: bold;
-  margin-bottom: 2%;
-}
-
-.post-checkout {
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.post-checkout h3 {
-  color: green;
-}
-
-h5 {
-  margin: 5% 0 0 0;
-}
-
-.docs-card {
-  background-color: white;
-  margin-top: 5%;
-  box-shadow: 4px 4px 4px 4px silver;
-  padding: 1%;
-}
-
-#dropin-container {
-  width: 55%;
-}/*# sourceMappingURL=style2.css.map */
