--- views/checkout/ach-debit.ejs
@@ -1,109 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>ACH Debit</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta3/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-eOJMYsd53ii+scO/bJGFsiCZc+5NDVN2yr8+0RDqr0Ql0h+rP48ckxlpbzKgwra6" crossorigin="anonymous">
-    <link rel='stylesheet' href='/style2.css' type='text/css'>
-    <link rel="stylesheet" href="/hostedFields.css" type="text/css">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"
-        integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p"
-        crossorigin="anonymous"></script>
-</head>
-
-<body>
-
-    <div id="hostedFields" class="container-fluid mx-auto">
-        <form action="/ach-vault" method="post" class="row" id="achForm">
-            <div class='col-12 spacing'>
-                <label>ACCOUNT NUMBER</label>
-                <input type="text" id="account-number" name="accountNumber">
-            </div>
-            <div class='col-12 spacing'>
-                <label>ROUTING NUMBER</label>
-                <input type="text" id="routing-number" value="011000015" name="routingNumber">
-            </div>
-            <div class='col-12 spacing'>
-                <input hidden id='clientToken' value='<%= clientToken %>'>
-                <input hidden id='nonce' name="nonce">
-                <button id="submitButton" class="btn btn-success" type="submit">Save</button>
-            </div>
-        </form>
-        <hr>
-        <code>
-            By clicking ["Checkout"], I authorize Braintree, a service of PayPal, on behalf of [your business name here] (i) to verify my bank account information using bank information and consumer reports and (ii) to debit my bank account.
-        </code>
-    </div>
-
-    <script src="https://js.braintreegateway.com/web/3.85.3/js/client.min.js"></script>
-    <script src="https://js.braintreegateway.com/web/3.85.3/js/us-bank-account.min.js"></script>
-    <script>
-        const clientToken = document.getElementById('clientToken').value
-        const button = document.getElementById('submitButton')
-        const form = document.getElementById('achForm')
-
-        braintree.client.create({
-            authorization: clientToken
-        }, (clientErr, clientInstance) => {
-            if (clientErr) {
-                console.error('There was an error creating the Client.');
-                throw clientErr;
-            }
-
-            braintree.usBankAccount.create({
-                client: clientInstance
-            }, (usBankAccountErr, usBankAccountInstance) => {
-                if (usBankAccountErr) {
-                    console.error('There was an error creating the USBankAccount instance.');
-                    throw usBankAccountErr;
-                }
-
-                button.addEventListener('click', () => {
-                    event.preventDefault()
-
-                    const bankDetails = {
-                        accountNumber: document.getElementById('account-number').value,
-                        routingNumber: document.getElementById('routing-number').value,
-                        accountType: "checking",
-                        ownershipType: "personal",
-                        billingAddress: {
-                            streetAddress: "123 Main St.",
-                            extendedAddress: "",
-                            locality: "Chicago",
-                            region: "IL",
-                            postalCode: "60615"
-                        }
-                    };
-
-                    if (bankDetails.ownershipType === 'personal') {
-                        bankDetails.firstName = "Tess";
-                        bankDetails.lastName = "Data";
-                    } else {
-                        bankDetails.businessName = "Company Y";
-                    }
-
-                    usBankAccountInstance.tokenize({
-                        bankDetails: bankDetails,
-                        mandateText: 'By clicking ["Checkout"], I authorize Braintree, a service of PayPal, on behalf of [your business name here] (i) to verify my bank account information using bank information and consumer reports and (ii) to debit my bank account.'
-                    }, (tokenizeErr, tokenizedPayload) => {
-                        if (tokenizeErr) {
-                            console.error('There was an error tokenizing the bank details.');
-                            throw tokenizeErr;
-                        }
-                        console.log(tokenizedPayload)
-                        document.getElementById("nonce").value = tokenizedPayload.nonce
-                        form.submit()
-                    });
-                })
-
-            });
-        });
-
-    </script>
-</body>
-
-</html>
