--- views/checkout/drop-in.ejs
@@ -1,119 +1,0 @@
-<!DOCTYPE html>
-<html lang='en'>
-
-<head>
-    <meta charset='UTF-8'>
-    <meta http-equiv='X-UA-Compatible' content='IE=edge'>
-    <meta name='viewport' content='width=device-width, initial-scale=1.0'>
-    <title>Drop-in UI</title>
-    <link href='https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta3/dist/css/bootstrap.min.css' rel='stylesheet'
-        integrity='sha384-eOJMYsd53ii+scO/bJGFsiCZc+5NDVN2yr8+0RDqr0Ql0h+rP48ckxlpbzKgwra6' crossorigin='anonymous'>
-    <link rel='stylesheet' href='/style2.css' type='text/css'>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
-</head>
-
-<body>
-    <div id='checkout-page' class='mx-auto'>
-        <form id='paymentForm' class='row g-3' action='/transaction-sale' method='post'>
-            <div class='col-12'>
-                <label for='amount'>Transaction Amount</label>
-                <br />
-                <input id='amount' type='number' name='amount'>
-            </div>
-            <hr />
-            <div style="display: flex; align-items: center; justify-content: center;">
-                <div id='dropin-container'></div>
-            </div>
-            <div class='col-12'>
-                <button hidden disabled id='submitButton' class='btn btn-success'>Purchase</button>
-            </div>
-            <input hidden id='clientToken' value='<%= clientToken %>'>
-            <input type='hidden' id='paymentMethodNonce' name='paymentMethodNonce' />
-            <input type='hidden' id='deviceData' name='deviceData' />
-        </form>
-        <div class='docs-card'>
-            <table class='table'>
-                <tbody>
-                    <tr>
-                        <th>Developer Docs</th>
-                        <td><a target='_blank'
-                                href='https://developer.paypal.com/braintree/docs/guides/drop-in/setup-and-integration'>developer.paypal.com/braintree/docs/guides/drop-in/setup-and-integration</a>
-                        </td>
-                    </tr>
-                    <tr>
-                        <th>Github.io</th>
-                        <td><a target='_blank'
-                                href='https://braintree.github.io/braintree-web-drop-in/docs/current/Dropin.html'>braintree.github.io/braintree-web-drop-in/docs/current/Dropin.html</a>
-                        </td>
-                    </tr>
-                    <tr>
-                        <th>GitHub</th>
-                        <td><a target='_blank'
-                                href='https://github.com/braintree/braintree-web-drop-in/blob/main/CHANGELOG.md'>github.com/braintree/braintree-web-drop-in/blob/main/CHANGELOG.md</a>
-                        </td>
-                    </tr>
-                    <tr>
-                        <th>Testing Card Values</th>
-                        <td><a target='_blank'
-                                href='https://developer.paypal.com/braintree/docs/reference/general/testing/node#credit-card-numbers'>developer.paypal.com/braintree/docs/reference/general/testing/node#credit-card-numbers</a>
-                        </td>
-                    </tr>
-                </tbody>
-            </table>
-        </div>
-        <script src="https://js.braintreegateway.com/web/dropin/1.40.2/js/dropin.min.js"></script>
-        <script type='text/javascript'>
-
-            const clientToken = document.getElementById('clientToken').value
-            const submitButton = document.getElementById('submitButton');
-            const form = document.getElementById('paymentForm')
-
-            console.log(`Client token is \n ${clientToken}`)
-
-            braintree.dropin.create({
-                authorization: clientToken,
-                card: {
-                    cardholderName: {
-                        required: false
-                    }
-                },
-                  paypal: {
-                    flow: "vault",
-                    amount: 45,
-                    currency: "USD"
-                  },      
-                venmo: {
-                    allowNewBrowserTab: true
-                },
-                container: '#dropin-container',
-                vaultManager: false,
-            }, (err, dropinInstance) => {
-                if (err) {
-                    console.error(err);
-                    return;
-                }
-
-                submitButton.removeAttribute('hidden')
-                submitButton.removeAttribute('disabled')
-
-                submitButton.addEventListener('click', (event) => {
-                    event.preventDefault()
-                    submitButton.setAttribute('disabled', true)
-                    dropinInstance.requestPaymentMethod((err, payload) => {
-                        if (err) {
-                            console.error(err);
-                            submitButton.removeAttribute('disabled')
-                            return
-                        }
-                        document.getElementById('paymentMethodNonce').value = payload.nonce;
-                        console.log(payload)
-                        console.log(payload.details)
-                        form.submit()
-                    });
-                });
-
-            });
-        </script>
-</body>
-
-</html>
