--- views/checkout/home.ejs
@@ -1,46 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Integration Testing</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-    <link rel="stylesheet" href="/style.css" type="text/css">
-</head>
-
-<body>
-    <div id="homePage" class="container-fluid">
-        <h1>Integration Testing</h1>
-        <main>
-            <section>
-                <a href="/drop-in" class="btn btn-primary">Drop-in UI</a>
-                <a href="/hosted-fields" class="btn btn-primary">Hosted Fields</a>
-                <a href="/3DS-drop-in" class="btn btn-primary">3D Secure (Drop-in UI)</a>
-                <a href="/3DS-hosted-fields" class="btn btn-primary">3D Secure (Hosted Fields)</a>
-                <a href="/3DS-vaulted" class="btn btn-primary">3D Secure (Vaulted Checkout)</a>
-                <a href="/paypal" class="btn btn-primary">PayPal</a>
-                <a href="/google-pay" class="btn btn-primary">Google Pay</a>
-                <a href="/local-payment-methods" class="btn btn-primary">Local Payment Methods</a>
-                <a href="/ach-debit" class="btn btn-primary">ACH Debit</a>
-                <a href="/payment-request-api" class="btn btn-primary">Payment Request API</a>
-            </section>
-            <hr />
-            <section>
-                <a href="#" class="btn btn-dark">Transactions</a>
-                <a href="#" class="btn btn-dark">Subscriptions</a>
-                <a href="#" class="btn btn-dark">Customers</a>
-                <a href="#" class="btn btn-dark">Payment Methods</a>
-                <a href="#" class="btn btn-dark">Payment Method Nonces</a>
-            </section>
-            <hr />
-            <section>
-                <a href="/import-subscriptions" class="btn btn-success">Import Subscriptions</a>
-            </section>
-        </main>
-    </div>
-</body>
-
-</html>
