--- views/credentials.ejs
@@ -1,56 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<%- include('partials/head.ejs') %>
-
-<body>
-  <div id="homePage" class="container-fluid">
-    <h1>Braintree Support Integration Training</h1>
-    <a id="homeButton" href="/"><i class="fa fa-home"></i></a>
-    <hr />
-    <main>
-      <% if (feedback.status==="SUCCESS" ) { %>
-        <h2>
-          You've successfully linked your sandbox! &nbsp;<i class="fa fa-check"
-            style="font-size: 48px; color: green"></i>
-        </h2>
-        <h5>
-          Go to your&nbsp;<a target="_blank"
-            href="https://sandbox.braintreegateway.com/merchants/<%=feedback.data.merchantId%>/home">Sandbox Control
-            Panel</a>
-          &nbsp;or proceed to the&nbsp;<a href="/client-token">next task.</a>
-        </h5>
-        <% } %>
-
-          <% if (feedback.status==="FAILURE" ) { %>
-            <h2>
-              <%= feedback.data.name %>
-            </h2>
-
-            <% if (feedback.data.name==="authenticationError" ) { %>
-              <h5>
-                <a target="_blank"
-                  href="https://developer.paypal.com/braintree/docs/reference/general/exceptions/node#authentication-error">This
-                  error</a>&nbsp; is raised when your API keys are incorrect.
-              </h5>
-              <h5>
-                Double check you've&nbsp;<a target="_blank"
-                  href="https://developer.paypal.com/braintree/docs/start/hello-server/node#install-and-configure">entered
-                  the correct API credentials</a>&nbsp;in the config gateway file.
-              </h5>
-              <% } %>
-
-                <% if (feedback.data.name==="TypeError" ) { %>
-                  <h5>
-                    Please enter your&nbsp;<a target="_blank"
-                      href="https://developer.paypal.com/braintree/docs/start/hello-server/node#install-and-configure">API
-                      credentials</a>&nbsp;in the config gateway file.
-                  </h5>
-                  <% } %>
-
-                    <% } %>
-    </main>
-  </div>
-</body>
-
-</html>
