+++ views/home.ejs
@@ -1,0 +1,39 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Braintree Support Integration Training</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
+    <link rel="stylesheet" href="/style.css" type="text/css">
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Raleway&display=swap" rel="stylesheet">
+</head>
+
+<body>
+    <div id="homePage" class="container-fluid">
+        <h1>Braintree Support Integration Training</h1>
+        <hr>
+        <main>
+            <section>
+                <a href="/drop-in" class="btn btn-success">Drop-in UI</a>
+                <a href="/hosted-fields" class="btn btn-success">Hosted Fields</a>
+                <a href="/three-d-secure" class="btn btn-success">3D Secure</a>
+                <a href="/paypal" class="btn btn-success">PayPal</a>
+            </section>
+            <hr>
+            <section>
+                <a href="#" class="btn btn-success">Transaction API Calls</a>
+                <a href="#" class="btn btn-success">Subscription API Calls</a>
+                <a href="#" class="btn btn-success">Customer API Calls</a>
+                <a href="#" class="btn btn-success ">Payment Method API Calls</a>
+            </section>
+        </main>
+    </div>
+</body>
+
+</html>
