+++ views/hosted-fields.ejs
@@ -90,8 +90,8 @@
     </div>
   </div>
 
-  <script src="https://js.braintreegateway.com/web/3.97.1/js/client.min.js"></script>
-  <script src="https://js.braintreegateway.com/web/3.97.1/js/hosted-fields.min.js"></script>
+  <script src="https://js.braintreegateway.com/web/3.84.0/js/client.min.js"></script>
+  <script src="https://js.braintreegateway.com/web/3.84.0/js/hosted-fields.min.js"></script>
   <script>
 
     const clientToken = document.getElementById("clientToken").value
@@ -149,6 +149,9 @@
         }, false);
       });
     });
+
+
+    // <--- ERROR HANDLING ---> 
 
     const errorMessageContainer = document.getElementById('errorMessages')
