--- views/more.ejs
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Integration Testing</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
-    <link rel="stylesheet" href="/style2.css" type="text/css">
-</head>
-
-<body>
-    <div id="homePage">
-        <h1>Integration Testing</h1>
-        <main>
-            <section>
-                <a href="/checkout/drop-in" class="btn btn-primary">Drop-in UI</a>
-                <a href="/checkout/hosted-fields" class="btn btn-primary">Hosted Fields</a>
-                <a href="/checkout/3DS-drop-in" class="btn btn-primary">3D Secure (Drop-in UI)</a>
-                <a href="/checkout/paypal" class="btn btn-primary">PayPal</a>
-                <a href="/checkout/google-pay" class="btn btn-primary">Google Pay</a>
-                <a href="/checkout/local-payment-methods" class="btn btn-primary">Local Payment Methods</a>
-                <a href="/checkout/ach-debit" class="btn btn-primary">ACH Direct Debit</a>
-            </section>
-        </main>
-    </div>
-</body>
-
-</html>
