--- views/partials/info.ejs
@@ -1,30 +1,0 @@
-<div class='docs-card'>
-    <table class='table'>
-        <tbody>
-            <tr>
-                <th>Developer Docs</th>
-                <td><a target='_blank'
-                        href='https://developer.paypal.com/braintree/docs/guides/drop-in/setup-and-integration'>developer.paypal.com/braintree/docs/guides/drop-in/setup-and-integration</a>
-                </td>
-            </tr>
-            <tr>
-                <th>Github.io</th>
-                <td><a target='_blank'
-                        href='https://braintree.github.io/braintree-web-drop-in/docs/current/Dropin.html'>braintree.github.io/braintree-web-drop-in/docs/current/Dropin.html</a>
-                </td>
-            </tr>
-            <tr>
-                <th>GitHub</th>
-                <td><a target='_blank'
-                        href='https://github.com/braintree/braintree-web-drop-in/blob/main/CHANGELOG.md'>github.com/braintree/braintree-web-drop-in/blob/main/CHANGELOG.md</a>
-                </td>
-            </tr>
-            <tr>
-                <th>Testing Card Values</th>
-                <td><a target='_blank'
-                        href='https://developer.paypal.com/braintree/docs/reference/general/testing/node#credit-card-numbers'>developer.paypal.com/braintree/docs/reference/general/testing/node#credit-card-numbers</a>
-                </td>
-            </tr>
-        </tbody>
-    </table>
-</div>
